refactor(frontend): global/router -> router
