+++ packages/frontend/src/pages/settings/index.vue
@@ -37,7 +37,7 @@ import { clearCache } from '@/scripts/clear-cache.js';
 import { instance } from '@/instance.js';
 import { PageMetadata, definePageMetadata, provideMetadataReceiver } from '@/scripts/page-metadata.js';
 import * as os from '@/os.js';
-import { useRouter } from '@/global/router/supplier.js';
+import { useRouter } from '@/router/supplier.js';
 
 const indexInfo = {
 	title: i18n.ts.settings,
